% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccwrpwbe_functions.R
\name{covid_lag}
\alias{covid_lag}
\title{Covid Lag}
\usage{
covid_lag(data, offset = 0, n = 1L, CC = F)
}
\arguments{
\item{data}{Main data for Pt loma or Hyperion}

\item{offset}{Number of days to shift concentrations}

\item{n}{Number of days of differencing for running cases count to go to new cases}

\item{CC}{CC = TRUE to also difference concentration n days}
}
\value{
Df with Cases_Offset column recording the new case count for the given day
}
\description{
Covid Lag takes in Main 'data' which has an aggregate total of cases and will
calculate lag differences between cases over 'n' days. Covid lag will also take in an offset
value and shift the concentration measurements by 'offset' days to line up with the new
case count. For example if offset = 7 concentration measurements recorded on '2020-04-20'
will be shifted to align with new cases recorded on '2020-04-27'. In addition the CC
parameter can be set to TRUE to take the difference in concentrations as well.
}
\examples{
\dontrun{
need to attach data for example
}
}
