% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccwrpwbe_functions.R
\name{date_range}
\alias{date_range}
\title{Date_Range}
\usage{
date_range(data, start = NULL, end = NULL, weekday = NULL)
}
\arguments{
\item{data}{df or time series with Date column}

\item{start}{Character data type expressing date of form 'YYYY-MM-DD'}

\item{end}{Character data type of form 'YYYY-MM-DD'}

\item{weekday}{Character name of weekday e.g. 'Tuesday' or 'Tue'}
}
\value{
Returns df filtered to date range defined by inputs
}
\description{
Takes df or ts, start, and end date range and returns df filtered to observations
in that time period. Additionally, can take in weekday argument and will only return observations
occurring on that weekday. Also will work with abbreviated weekday such as 'Tue'. Function is also pipe
compatible with Tidy Verse
}
\examples{
data <- data.frame(Date= seq(as.Date("2021/04/20"), by = "day", length.out = 100), Obs = rnorm(100))

#Filter for all days between start and end date
date_range(data , start = '2021-05-10', end = '2021-06-15')

#Filter for every Tuesday between start and end
date_range(data , start = '2021-05-10', end = '2021-06-15',weekday = 'Tuesday')

#Example with pipe
data \%>\%



}
