% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccwrpwbe_functions.R
\name{ggcorr_offset}
\alias{ggcorr_offset}
\title{Title}
\usage{
ggcorr_offset(data, range, target, n = 1L, CC = F, ...)
}
\arguments{
\item{data}{Formatted Hyperion or Pt.Loma df}

\item{range}{Vector of offset values to shift concentrations}

\item{target}{Character 'N1' or 'N2'}

\item{n}{Number of days to difference}

\item{CC}{Boolean for if concentrations are to be differenced}

\item{...}{Parameters from \code{\link{date_range}}}
}
\value{
Prints plot with normalization method across rows and offsets across cols
}
\description{
Title
}
\examples{
#Plot offset concentrations and cases by 1,3,5,7 days
Hyperion \%>\% ggcorr_offset(c(1,3,5,7) , target = 'N1' , start = '2020-05-25' , end = '2020-10-20')
}
