% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccwrpwbe_functions.R
\name{ggimpute_eval}
\alias{ggimpute_eval}
\title{ggimpute_eval}
\usage{
ggimpute_eval(data, target, rate, seeds, type = "boxplot", ...)
}
\arguments{
\item{data}{result from impute_eval}

\item{target}{need to add}

\item{rate}{need to add}

\item{seeds}{need to add}

\item{type}{need to add}

\item{...}{need to add}
}
\value{

}
\description{
Need to add
}
