% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccwrpwbe_functions.R
\name{ggkeystone}
\alias{ggkeystone}
\title{ggkeystone}
\usage{
ggkeystone(data, target, k = 7, lag = 1L, offset = 0, CC = F, ...)
}
\arguments{
\item{data}{Formatted Hyperion or Pt.Loma df}

\item{target}{Character 'N1' or 'N2' "assays"}

\item{k}{Number of days for rolling average.}

\item{lag}{Number of days for lagged difference from n parameter in \code{\link{covid_lag}}}

\item{offset}{Number of days to shift concentrations from \code{covid_lag}}

\item{CC}{CC parameter from \code{covid_lag}}

\item{...}{\code{\link{date_range}}}
}
\value{
Returns keystone plot across all 4 normalization methods
}
\description{
The ggkeystone function takes in arguments from covid_lag as well as
date_range. It also takes in a parameter k which represents the number of days to take a rolling
average over. It returns the "keystone" graphic showing the relationship between waste water concentration
and new Covid cases.
}
\examples{
#note should be used with some imputation method such as na_kalman()
library(imputeTS)
Hyperion \%>\% na_kalman() \%>\% ggkeystone(target = "N1") + ylab("")
}
