% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccwrpwbe_functions.R
\name{ggplot_impute}
\alias{ggplot_impute}
\title{ggimpute}
\usage{
ggplot_impute(data, type = "boxplot", target_name = NULL)
}
\arguments{
\item{data}{df with the results of impute_analysis}

\item{type}{Type of plot to show, either boxplot or error bar}

\item{target_name}{Which Assay?}
}
\value{

}
\description{
The function ggimpute takes in a df from \code{\link{impute_analysis}} a type of plot
and returns a boxplot or and errorbard plot showing the imputation analysis results. This function is used in
\code{\link{ggimpute_eval}} to show the results across multiple normalization methods.
}
