% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperion_data.R
\docType{data}
\name{Hyperion}
\alias{Hyperion}
\title{Hyperion Data}
\format{
A \code{df} with 18 columns

\describe{
\item{Date}{date of observation}
\item{BoCoV}{measured BoCoV}
\item{PMMV}{measured PMMV}
\item{N1_Copies_100mL_adj}{unnormalized N1}
\item{N2_Copies_100mL_adj}{unnormalized N2}
\item{Navg_Copies_100mL_adj}{unnormalized Navg}
\item{N1_Norm_BP}{N1 normalized with Bocov+PMMV}
\item{N2_Norm_BP}{N2 normalized with Bocov+PMMV}
\item{Navg_Norm_BP}{NAVG normalized with Bocov+PMMV}
\item{N1_Norm_B}{N1 nomralized with Bocov}
\item{N2_Norm_B}{N2 normalized with Bocov}
\item{Navg_Norm_B}{Navg normalized with Bocov}
\item{N1_Norm_P}{N1 normalized with PMMV}
\item{N2_Norm_P}{N2 normalized with PMMV}
\item{Navg_Norm_P}{Navg Normalized with PMMV}
\item{htp_cases}{Cases in Hyperion territory from LA Times Github}
\item{cases}{Cases in Hyperion territory from LA Times Github}
\item{icu_positive_patients}{Running total of ICU postive patients}
}
}
\usage{
Hyperion
}
\description{
Hyperion treatment plant dataset which includes concentration observations
for all normalization methods on N1/N2 assays as well as Bocov measurements, PMMOV measurements, and
a running total of ICU positive patients. The date range covered is 2020-01-26 to 2021-04-20. Although
waste water testing does not begin until 2020-04-20. Case and ICU counts were collected from LA Times
Github using Hyperion service cities.
}
\keyword{datasets}
