% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccwrpwbe_functions.R
\name{laglead}
\alias{laglead}
\title{LagLead}
\usage{
laglead(x, offset)
}
\arguments{
\item{x}{Data}

\item{offset}{Number of intervals index data is to be shifted}
}
\value{
For offset > 0, data with index shifted forward by offset.
For Offset < 0, data shifted back by offset
}
\description{
Runs \code{\link[dplyr]{lag}} or \code{\link[dplyr]{lead}} functions
depending on if offset is positive or negative. Positive values of offset
shift data forward. Negative values shift data backwards.
}
\examples{
#Shift data forward by offset = 2
data <- data.frame('A' = runif(n=25,min=50,max=100),'B'=runif(n=25,min=50,max=100))
laglead(x = data , 2)

#Shift data backward by offset = -2
laglead(x = data , -2)
}
