% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccwrpwbe_functions.R
\name{minmax}
\alias{minmax}
\title{Minmax Scale}
\usage{
minmax(x)
}
\arguments{
\item{\code{x}}{numeric vector}
}
\value{
\code{x} values from input vector scaled to be between 0 and 1
}
\description{
Scales values of numeric vector to be between 0 and 1
}
\examples{
x <- runif(10, min =0 ,max=100)
minmax(x)


}
