% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccwrpwbe_functions.R
\name{date_range}
\alias{date_range}
\title{Date_Range}
\usage{
date_range(data, start = NULL, end = NULL, weekday = NULL)
}
\arguments{
\item{data}{df or time series with Date column}

\item{start}{Character data type expressing date of form 'YYYY-MM-DD'}

\item{end}{Character data type of form 'YYYY-MM-DD'}

\item{weekday}{Character name of weekday e.g. 'Tuesday' or 'Tue'}
}
\value{
Returns df filtered to date range defined by inputs
}
\description{
Takes in df or ts object, start, and end date range and returns df filtered to observations
in that time period. Additionally, can take in weekday argument and will only return observations
occurring on that weekday. Also will work with abbreviated weekday such as 'Tue'. Function is also pipe
compatible with Tidy Verse
}
\examples{
#Example For All Days
Hyperion \%>\% date_range(start='2020-04-20', end = '2020-07-11')

#Example for selecting a weekday only
Hyperion \%>\% date_range(start='2020-04-20', end = '2020-07-11', weekday='Tuesday')

#Example with abbreviated day
Hyperion \%>\% date_range(start='2020-04-20', end = '2020-07-11', weekday='Tue')

}
